Move Hash256 and EthereumAddress to the types package
