+++ packages/backend/src/core/StateUpdateCollector.ts
@@ -1,8 +1,7 @@
 import { decodeOnChainData as _decodeOnChainData } from '@explorer/encoding'
 import { RollupState } from '@explorer/state'
-import { PedersenHash } from '@explorer/types'
+import { Hash256, PedersenHash } from '@explorer/types'
 
-import { Hash256 } from '../model'
 import { PageRepository } from '../peripherals/database/PageRepository'
 import { RollupStateRepository } from '../peripherals/database/RollupStateRepository'
 import { StateTransitionFactRecord } from '../peripherals/database/StateTransitionFactsRepository'
