+++ packages/backend/src/core/sync/BlockDownloader.ts
@@ -1,7 +1,6 @@
-import { json } from '@explorer/types'
+import { Hash256, json } from '@explorer/types'
 import { providers } from 'ethers'
 
-import { Hash256 } from '../../model'
 import {
   BlockRecord,
   BlockRepository,
