+++ packages/backend/src/model/index.ts
@@ -1,3 +1,1 @@
 export * from './BlockRange'
-export * from './EthereumAddress'
-export * from './Hash256'
