+++ packages/backend/test/peripherals/database/FactToPageRepository.test.ts
@@ -1,6 +1,6 @@
+import { Hash256 } from '@explorer/types'
 import { expect } from 'earljs'
 
-import { Hash256 } from '../../../src/model'
 import {
   FactToPageRecord,
   FactToPageRepository,
