+++ packages/backend/test/peripherals/database/StateTransitionFactsRepository.test.ts
@@ -1,6 +1,6 @@
+import { Hash256 } from '@explorer/types'
 import { expect } from 'earljs'
 
-import { Hash256 } from '../../../src/model'
 import {
   StateTransitionFactRecord,
   StateTransitionFactRepository,
