+++ packages/types/src/index.ts
@@ -1,3 +1,5 @@
 export * from './AssetId'
+export * from './EthereumAddress'
+export * from './Hash256'
 export * from './json'
 export * from './PedersenHash'
