+++ packages/types/test/EthereumAddress.test.ts
@@ -1,6 +1,6 @@
 import { expect } from 'earljs'
 
-import { EthereumAddress } from '../../src/model'
+import { EthereumAddress } from '../src'
 
 describe('EthereumAddress', () => {
   it('accepts lowercase addresses', () => {
